Nah, this is going to make it too hard to track the history of the Evaluator class
