+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/FunctionalEvaluator.java
@@ -2,11 +2,16 @@ package sa.com.cloudsolutions.antikythera.evaluator;
 
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.body.CallableDeclaration;
+import com.github.javaparser.ast.body.MethodDeclaration;
 import com.github.javaparser.ast.expr.Expression;
+import sa.com.cloudsolutions.antikythera.exception.EvaluatorException;
+import sa.com.cloudsolutions.antikythera.parser.MCEWrapper;
 
 import java.util.Map;
 
 public class FunctionalEvaluator extends AbstractEvaluator implements ExpressionEvaluator{
+    private MethodDeclaration methodDeclaration;
+
     public FunctionalEvaluator(String className) {
         super(className);
     }
@@ -31,8 +36,25 @@ public class FunctionalEvaluator extends AbstractEvaluator implements Expression
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
+    public void setMethod(MethodDeclaration md) {
+        this.methodDeclaration = md;
+    }
+
+    /**
+     * Execute a method call.
+     * @param wrapper the method call expression wrapped so that the argument types are available
+     * @return the result of executing that code.
+     * @throws EvaluatorException if there is an error evaluating the method call or if the
+     *          feature is not yet implemented.
+     */
     @Override
-    public Variable executeMethod(CallableDeclaration<?> cd) throws ReflectiveOperationException {
-        return null;
+    public Variable executeMethod(MCEWrapper wrapper) throws ReflectiveOperationException {
+        returnFrom = null;
+
+        Variable v = executeMethod(methodDeclaration);
+        if (v != null && v.getValue() == null) {
+            v.setType(methodDeclaration.getType());
+        }
+        return v;
     }
 }
