+++ src/test/java/sa/com/cloudsolutions/antikythera/evaluator/Functional.java
@@ -9,13 +9,18 @@ public class Functional {
         System.out.println(f.apply("Ashfaloth"));
     }
 
-    private void greet() {
+    private void greet1() {
         printHello( a -> "hello " + a );
+    }
+
+    private void greet2() {
+        printHello( a -> { return "hello " + a; });
     }
 
     public static void main(String[] args) {
         Functional f = new Functional();
-        f.greet();
+        f.greet1();
+        f.greet2();
     }
 
 }
