+++ src/test/java/sa/com/cloudsolutions/antikythera/evaluator/TestFunctional.java
@@ -4,7 +4,8 @@ import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.body.MethodDeclaration;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import sa.com.cloudsolutions.antikythera.configuration.Settings;
 import sa.com.cloudsolutions.antikythera.depsolver.ClassProcessor;
 import sa.com.cloudsolutions.antikythera.exception.AntikytheraException;
@@ -28,10 +29,11 @@ public class TestFunctional extends TestHelper{
         System.setOut(new PrintStream(outContent));
     }
 
-    @Test
-    void testBiFunction() throws ReflectiveOperationException {
+    @ParameterizedTest
+    @ValueSource(strings = {"greet1", "greet2"})
+    void testBiFunction(String name) throws ReflectiveOperationException {
         CompilationUnit cu = compiler.getCompilationUnit();
-        MethodDeclaration method = cu.findFirst(MethodDeclaration.class, m -> m.getNameAsString().equals("greet")).orElseThrow();
+        MethodDeclaration method = cu.findFirst(MethodDeclaration.class, m -> m.getNameAsString().equals(name)).orElseThrow();
         Variable v = evaluator.executeMethod(method);
         assertNull(v.getValue());
         assertEquals("Hello Ashfaloth\n", outContent.toString());
